+++ src/uhostmode.pas
@@ -191,7 +191,7 @@ begin
               SetCallsign;
             end;
             if Length(Text) > 0 then
-              ChannelBuffer[Channel] := ChannelBuffer[Channel] + #27'[31m' + '>>> ERROR: ' + Text + #13#27'[0m';
+              ChannelBuffer[Channel] := ChannelBuffer[Channel] + #13#27'[31m' + '>>> ERROR: ' + Text + #27'[0m'#13;
           end;
         end;
         3: // Link Status
@@ -201,7 +201,7 @@ begin
             Text := ReceiveDataUntilZero;
             if Length(Text) > 0 then
             begin
-              ChannelBuffer[Channel] := ChannelBuffer[Channel] + #27'[32m' + '>>> LINK STATUS: ' + Text + #13#27'[0m';
+              ChannelBuffer[Channel] := ChannelBuffer[Channel] + #13#27'[32m' + '>>> LINK STATUS: ' + Text + #27'[0m'#13;
               LinkStatus := DecodeLinkStatus(Text);
               ChannelStatus[channel][6] := LinkStatus[0]; // Status Text CONNECTED, DISCONNECTED, etc
               ChannelStatus[channel][7] := LinkStatus[1]; // Call of the other station