+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - ADD: TFKISS Support. More Details in the README.
 - FIX: create default config if no config file exist.
 - FIX: tnc parity error.
+- FIX: Select TFKISS Socket file Dialog.
 
 ## v0.3.0